Fix BSM formula in MathUtilities. Add European Option analytic function.

+++ MathUtilities.cpp
@@ -19,7 +19,7 @@ double Covariance(const std::vector<double>& v1, const std::vector<double>& v2) 
     return x - Expectation(v1) * Expectation(v2);
 }
 
-double BlackScholesFormula(double S, double K, double r, double T, double sig, int iscall, int div, int isdivdiscrete) {
+double BlackScholesFormula(double S, double K, double r, double T, double sig, int iscall, double div, int isdivdiscrete) {
     double adj_S;
     if (isdivdiscrete) {
         // cash amount
